% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{get_Het}
\alias{get_Het}
\title{Information in Heterogeneity}
\usage{
get_Het(CountsMatrix)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}
}
\value{
Numeric vector of length equal to the number of features (rows in CountsMatrix)
}
\description{
Calculates Het, the information encoded in heterogeneity, in a gene-wise manner
}
\details{
Output has units of bits. Theoretical maximum of calculation is log number of cells.
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(2,3,3),
                               j = c(1,1,2),
                               x = c(1,1,1))
Counts
Het <-  get_Het(CountsMatrix = Counts)

}
