% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{get_HetMicro}
\alias{get_HetMicro}
\title{Calaculate Information not in Model}
\usage{
get_HetMicro(CountsMatrix, groups, GroupedCounts, reduced = T)
}
\arguments{
\item{CountsMatrix}{dgCMatrix}

\item{groups}{factor}

\item{GroupedCounts}{dgCMatrix}

\item{reduced}{logical}
}
\value{

}
\description{
Calaculate Information not in Model
}
