% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{getFreqShrink}
\alias{getFreqShrink}
\title{Calculate cell frequencies for shrinakge estimator}
\usage{
getFreqShrink(transposeCounts, ind, N, Total)
}
\arguments{
\item{transposeCounts}{Transposed sparse count matrix}

\item{ind}{Integer indicating chosen gene (row number in count matrix)}

\item{N}{Number of cells}

\item{Total}{Integer of total counts per cell}
}
\value{
Numeric vector of shrinkage cell frequencies
}
\description{
Calculate cell frequencies for shrinakge estimator
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(2,3,3),
                               j = c(1,1,2),
                               x = c(1,1,1))
N <- ncol(Counts)
Total <- Matrix::rowSums(Counts)
getFreqShrink(Matrix::t(Counts), 1, N, Total)

}
