% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{getHetMacro}
\alias{getHetMacro}
\title{Calaculate Information in Model}
\usage{
getHetMacro(CountsMatrix, Groups, base = 2)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{Groups}{Factor of cell identities}

\item{base}{base of log}
}
\value{
Numeric vector of gene-wise HetMacro
}
\description{
Calculates HetMacro, the heterogeneity explained by a given model of population structure,
in a gene-wise manner
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,1,1,1,2,2,2),
                               j = c(1,2,3,4,1,2,3),
                               x = c(2,2,2,2,3,3,2))
Ident <- factor(c("1", "1", "2", "2"))
getHetMacro(Counts, Ident)
}
