% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouped.R
\name{group_Counts}
\alias{group_Counts}
\title{Group counts by cell identity}
\usage{
group_Counts(CountsMatrix, groups)
}
\arguments{
\item{CountsMatrix}{dgCMatrix}

\item{groups}{factor}
}
\value{
dgCMatrix
}
\description{
For a given grouping of cells, produces a sparse matrix where each column is the total counts in each group.
}
\examples{

}
