% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency.R
\name{groupCounts}
\alias{groupCounts}
\title{Sum counts by group}
\usage{
groupCounts(CountsMatrix, Groups)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{Groups}{Factor of cell identities}
}
\value{
Feature x group matrix of class dgCMatrix
}
\description{
For a given grouping of cells, produces a sparse matrix where each column is the total counts for each group.
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,1,1,1,2,2,2),
                               j = c(1,2,3,4,1,2,3),
                               x = c(2,2,2,2,3,3,2))
Ident <- factor(c("1", "1", "2", "2"))
groupCounts(Counts, Ident)
}
