% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{regularisedTotalHetMacro}
\alias{regularisedTotalHetMacro}
\title{Calculate regularised total information explained}
\usage{
regularisedTotalHetMacro(CountsMatrix, Groups, shrinkage = T, lambda)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{Groups}{Factor of cell identities}

\item{shrinkage}{Boolean flag on whether to use James-Stein type shrinkage estimator}

\item{lambda}{Scale factor of regularisation; defaults to ratio of total information obtained to theoretical maximum}
}
\value{
Regularised total information explained
}
\description{
Calculates total macro-heterogeneity regularised by the entropy of the group structure
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,1,1,1,2,2,2),
                               j = c(1,2,3,4,1,2,3),
                               x = c(2,2,2,2,3,3,2))
Ident <- factor(c("1", "1", "2", "2"))
regularisedTotalHetMacro(Counts, Ident)
}
