% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomSimulation.R
\name{simulate_Hom}
\alias{simulate_Hom}
\title{Simualtion of Homogenous gene expression}
\usage{
simulate_Hom(CountsMatrix, TotalCounts, NumTrials, depth_adjusted = T)
}
\arguments{
\item{CountsMatrix}{dgCMatrix}

\item{TotalCounts}{vector}

\item{NumTrials}{integer}

\item{depth_adjusted}{logical}
}
\value{

}
\description{
Simualtion of Homogenous gene expression
}
