% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomSimulation.R
\name{simulateHom}
\alias{simulateHom}
\title{Simulate Homogenous gene expression}
\usage{
simulateHom(
  CountsMatrix,
  NumTrials = 50,
  SimStep = 0.2,
  TotalCounts = NA,
  DepthAdjusted = T
)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{NumTrials}{Number of trials of simulation}

\item{SimStep}{Step size, in log10 counts, between number of counts simulated}

\item{TotalCounts}{Optional integer vector of total counts for simualted cells.
Default set internally is 10 to max cell counts with SimStep steps}

\item{DepthAdjusted}{Logical flag for whether simulated cells should be
assigned counts in ratio of corresponding library sizes}
}
\value{

}
\description{
Simulates null model of gene expression for given count matrix, producing
values for Het expected for each gene under the null.
}
\details{
Null model is simulated over set of cells of size N, where N is the number of cells in CountsMatrix.
Some number of counts is randomly assigned to cells either with equal probability or in proportion to observed count depths (total counts per cell)
Het is calculated for each trial. The mean Het for each number of counts is found.
The null Het for each gene is then assigned by interpolation.
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,1,2,2),
                               j = c(1,2,1,2),
                               x = c(1000,9000,5000,5000))
SimCounts <- 10000
simulateHom(Counts)
}
