% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{subtract_HetSparse}
\alias{subtract_HetSparse}
\title{Subtract information due to count sparsity}
\usage{
subtract_HetSparse(Het, CountsMatrix)
}
\arguments{
\item{Het}{vector}

\item{CountsMatrix}{dgCMatrix}
}
\value{

}
\description{
Subtract information due to count sparsity
}
