% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{subtractHetSparse}
\alias{subtractHetSparse}
\title{Subtract information due to count sparsity}
\usage{
subtractHetSparse(CountsMatrix, Het)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{Het}{Numeric vector of Het values of length equal to number of features}
}
\value{
Numeric vector of length equal to the number of features (rows in CountsMatrix)
}
\description{
Subtract information due to count sparsity
}
\details{
Sparsity is defined here as a feature having some number of counts M, less than
N, the number of cells.
The minimum information of a sparse feature is log N/M bits, which is
subtracted from Het.
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,2,2,2),
                               j = c(1,1,2,2),
                               x = rep(1, 4))
Het <- getHet(Counts)
subtractHetSparse(Counts, Het)
}
