% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{get_Het}
\alias{get_Het}
\title{Calculate Information in Heterogeneity}
\usage{
get_Het(CountsMatrix)
}
\arguments{
\item{CountsMatrix}{dgCMatrix}
}
\value{

}
\description{
Calculate Information in Heterogeneity
}
