% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{get_Het}
\alias{get_Het}
\title{Information in Heterogeneity}
\usage{
get_Het(CountsMatrix)
}
\arguments{
\item{CountsMatrix}{dgCMatrix. Feature x cell sparse counts matrix.}
}
\value{
Numeric vector of length equal to the number of features (rows in CountsMatrix)
}
\description{
Calculates Het, the information encoded in heterogeneity, on a gene-wise basis.
}
\details{
Output units of bits. Theoretical maximum of calculation is log number of cells.
}
