% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{get_HetMacro}
\alias{get_HetMacro}
\title{Calaculate Information in Model}
\usage{
get_HetMacro(CountsMatrix, groups, GroupedCounts)
}
\arguments{
\item{CountsMatrix}{dgCMatrix}

\item{groups}{factor}

\item{GroupedCounts}{dgCMatrix}
}
\value{

}
\description{
Calaculate Information in Model
}
