% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{getHetMicro}
\alias{getHetMicro}
\title{Calaculate Information not in Model}
\usage{
getHetMicro(CountsMatrix, Groups, full = F, subtractSparsity = F)
}
\arguments{
\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}

\item{Groups}{Factor of cell identities}

\item{full}{Logical flag for whether to return just HetMicro or full Het by group}

\item{subtractSparsity}{Subtract information due to count sparsity. If full also TRUE, also applies to each group.}
}
\value{

}
\description{
Calculates HetMicro, the heterogeneity within each group of cells, or its weigthed
average (default)
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,1,1,1,2,2,2),
                               j = c(1,2,3,4,1,2,3),
                               x = c(2,2,2,2,3,3,2))
Ident <- factor(c("1", "1", "2", "2"))
GroupCounts <- groupCounts(Counts, Ident)
getHetMicro(Counts, Ident)
}
