% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Het.R
\name{subtract_HetSparse}
\alias{subtract_HetSparse}
\title{Subtract information due to count sparsity}
\usage{
subtract_HetSparse(Het, CountsMatrix)
}
\arguments{
\item{Het}{Numeric vector of Het values of length equal to number of features}

\item{CountsMatrix}{Feature x cell sparse counts matrix of class dgCMatrix}
}
\value{
Numeric vector of length equal to the number of features (rows in CountsMatrix)
}
\description{
Subtract information due to count sparsity
}
\details{
Sparisty is defined here as a feature having some number of counts M, less than
N, the number of cells.
The minimum information of a sparse feature is log N/M bits, which is
subtracted from Het.
}
\examples{
Counts <- Matrix::sparseMatrix(i = c(1,2,2,2),
                               j = c(1,1,2,2),
                               x = rep(1, 4))
Het <- get_Het(Counts)
subtract_HetSparse(Het, Counts)
}
